+++ src/test/java/com/anagha/DivisionTest.java
@@ -1,10 +1,10 @@
 package com.anagha;
 import org.junit.Test;
-import com.anagha.division;
+import com.anagha.Division;
 public class DivisionTest {
     @Test
     public void testing() {
-    int a=200,b=74;
+    float a=200f,b=74f;
     Division d= new Division();
     float result=d.div(a,b);
     System.out.println(result);
